requirements.txt
numpy
torch
transformers
Pillow
# blender-python
opencv-python
pytest